Add preliminary backlight code to matrix
During matrix_init we try to find the caps and sym lock keys, and we
save their locations for later.
During matrix_scan we also scan through the LED array. Conveniently, the
number of keyboard columns is 3X the number of rows, so we just switch
to a new LED row every 3 columns.